Add Flask module for requiring membership to see CVu.
The idea is that we configure Apache to direct /login and /logout, plus
all pages under /journal/cvu to Flask. It will ensure all access to
/journal/cvu/* latter are logged in and are ACCU members; if not, they
are redirected to the login page.

I have tested this standalone, but not in the Apache setup, so this
might not fly.

Also, I've gone with the Python import style the Flask docs use, though
I don't like from ... import.

To avoid including the database password in the source, this requires a
configuration /etc/accu/website.cfg (or ~<apache user>/.accu.cfg)
containing the location of the MySQL database and the password for the
ACCU Xaraya user. The location defaults to 'localhost', the password to
in incorrect value.

[database]
host=localhost
password=<password>