Avoid unnecessary \n in adoc output.
In places where we have stuff that needs to be at the start of a line,
or to have a blank line before it, only put a newline in if the current
output doesn't end in one.

This will be important soon when we try to ensure list items with blocks
in them get continuation markers before the block.